ec/google/chromeec: Update google_chromeec_get_board_version prototype
The helper function to get the board version from EC returns 0 on
failure. But 0 is also a valid board version. Update the helper function
to return -1 on failure and update the use-cases.

BUG=b:114001972,b:114677884,b:114677887

Tested-by: Karthikeyan Ramasubramanian <kramasub@google.com>
Reviewed-on: https://review.coreboot.org/28576
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/mainboard/google/zoombini/variants/baseboard/include/baseboard/variants.h
@@ -23,7 +23,7 @@
 #include <vendorcode/google/chromeos/chromeos.h>
 
 /* Return the board id for the current variant board. */
-uint8_t variant_board_id(void);
+int variant_board_id(void);
 
 /*
  * The next set of functions return the gpio table and fill in the number of
